+++ pmd-kotlin/src/main/java/net/sourceforge/pmd/lang/kotlin/KotlinHandler.java
@@ -12,13 +12,31 @@ import net.sourceforge.pmd.lang.ast.Parser;
 import net.sourceforge.pmd.lang.kotlin.ast.InternalApiBridge;
 import net.sourceforge.pmd.lang.kotlin.ast.PmdKotlinParser;
 import net.sourceforge.pmd.lang.kotlin.internal.KotlinDesignerBindings;
+import net.sourceforge.pmd.lang.kotlin.rule.xpath.internal.KotlinHasAnnotationFunction;
+import net.sourceforge.pmd.lang.kotlin.rule.xpath.internal.KotlinHasUnresolvedReferenceFunction;
+import net.sourceforge.pmd.lang.kotlin.rule.xpath.internal.KotlinIsNullableFunction;
+import net.sourceforge.pmd.lang.kotlin.rule.xpath.internal.KotlinMatchesSigFunction;
+import net.sourceforge.pmd.lang.kotlin.rule.xpath.internal.KotlinModifiersFunction;
+import net.sourceforge.pmd.lang.kotlin.rule.xpath.internal.KotlinTypeIsExactlyFunction;
+import net.sourceforge.pmd.lang.kotlin.rule.xpath.internal.KotlinTypeIsFunction;
 import net.sourceforge.pmd.lang.rule.xpath.impl.XPathHandler;
 import net.sourceforge.pmd.util.designerbindings.DesignerBindings;
 
-
 public class KotlinHandler extends AbstractPmdLanguageVersionHandler {
 
-    private static final XPathHandler XPATH_HANDLER = XPathHandler.noFunctionDefinitions();
+    // Holder-class pattern: defers initialization until getXPathHandler() is first called,
+    // by which time KotlinLanguageModule is fully registered in LanguageRegistry.
+    private static final class XPathHandlerHolder {
+        static final XPathHandler HANDLER =
+                XPathHandler.getHandlerForFunctionDefs(
+                        KotlinTypeIsFunction.INSTANCE,
+                        KotlinTypeIsExactlyFunction.INSTANCE,
+                        KotlinMatchesSigFunction.INSTANCE,
+                        KotlinHasAnnotationFunction.INSTANCE,
+                        KotlinIsNullableFunction.INSTANCE,
+                        KotlinHasUnresolvedReferenceFunction.INSTANCE,
+                        KotlinModifiersFunction.INSTANCE);
+    }
 
     private PmdKotlinParser parser;
 
@@ -38,7 +56,7 @@ public class KotlinHandler extends AbstractPmdLanguageVersionHandler {
 
     @Override
     public XPathHandler getXPathHandler() {
-        return XPATH_HANDLER;
+        return XPathHandlerHolder.HANDLER;
     }
 
     @Override