+++ pmd-kotlin/src/main/java/net/sourceforge/pmd/lang/kotlin/ast/KtFunctionValueParameterAttributes.java
@@ -1,0 +1,18 @@
+/*
+ * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
+ */
+
+package net.sourceforge.pmd.lang.kotlin.ast;
+
+import net.sourceforge.pmd.annotation.Experimental;
+
+/**
+ * @since 7.27.0
+ * @experimental See {@link AttributeView}.
+ */
+@Experimental
+public class KtFunctionValueParameterAttributes extends AttributeView<KotlinParser.KtFunctionValueParameter> implements HasTypeName {
+    public KtFunctionValueParameterAttributes(KotlinParser.KtFunctionValueParameter node) {
+        super(node);
+    }
+}