+++ pmd-kotlin/src/test/resources/net/sourceforge/pmd/lang/kotlin/rule/xpath/typeIs/TypeAnnotationAttributes.kt
@@ -1,0 +1,16 @@
+package nl.stokpop.test
+
+import java.util.Calendar
+
+@Deprecated("use Foo instead")
+class TypeAnnotationAttributes {
+
+    @Deprecated("use newProcess instead")
+    fun process(cal: Calendar): String {
+        try {
+            return cal.time.toString()
+        } catch (e: IllegalArgumentException) {
+            return "error"
+        }
+    }
+}